+++ errors/errors.js
@@ -1,6 +1,12 @@
+// коды ошибок
+
 module.exports.INCORRECT_DATA_ERROR_CODE = 400;
 module.exports.NOT_FOUND_ERROR_CODE = 404;
 module.exports.USER_EXISTS_ERROR_CODE = 409;
 module.exports.UNAUTHORIZED_ERROR_CODE = 401;
 module.exports.FORBIDDEN_DELETE_ERROR_CODE = 403;
 module.exports.DEFAULT_ERROR_CODE = 500;
+
+// Тексты сообщений об ошибках
+
+module.exports.BAD_URL_FORMAT_MESSAGE = 'Указан некорректный URL';
